+++ crates/portfolio-manager/src/models.rs
@@ -1,12 +1,14 @@
-use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
+use rust_decimal::Decimal;
 
 #[derive(Debug, Clone, Serialize, Deserialize, sqlx::FromRow)]
 pub struct Position {
     pub id: Option<i64>,
     pub symbol: String,
-    pub shares: f64,
-    pub entry_price: f64,
+    #[sqlx(try_from = "f64")]
+    pub shares: Decimal,
+    #[sqlx(try_from = "f64")]
+    pub entry_price: Decimal,
     pub entry_date: String,
     pub notes: Option<String>,
     pub created_at: Option<String>,
@@ -16,10 +18,10 @@ pub struct Position {
 pub struct PositionWithPnL {
     #[serde(flatten)]
     pub position: Position,
-    pub current_price: f64,
-    pub market_value: f64,
-    pub cost_basis: f64,
-    pub unrealized_pnl: f64,
+    pub current_price: Decimal,
+    pub market_value: Decimal,
+    pub cost_basis: Decimal,
+    pub unrealized_pnl: Decimal,
     pub unrealized_pnl_percent: f64,
 }
 
@@ -28,10 +30,13 @@ pub struct Trade {
     pub id: Option<i64>,
     pub symbol: String,
     pub action: String, // "buy" or "sell"
-    pub shares: f64,
-    pub price: f64,
+    #[sqlx(try_from = "f64")]
+    pub shares: Decimal,
+    #[sqlx(try_from = "f64")]
+    pub price: Decimal,
     pub trade_date: String,
-    pub commission: f64,
+    #[sqlx(try_from = "f64")]
+    pub commission: Decimal,
     pub notes: Option<String>,
     pub profit_loss: Option<f64>,
     pub profit_loss_percent: Option<f64>,
@@ -42,10 +47,10 @@ pub struct Trade {
 pub struct TradeInput {
     pub symbol: String,
     pub action: String,
-    pub shares: f64,
-    pub price: f64,
+    pub shares: Decimal,
+    pub price: Decimal,
     pub trade_date: String,
-    pub commission: Option<f64>,
+    pub commission: Option<Decimal>,
     pub notes: Option<String>,
 }
 
@@ -72,9 +77,9 @@ pub struct AlertInput {
     pub alert_type: String,
     pub signal: String,
     pub confidence: f64,
-    pub current_price: Option<f64>,
-    pub target_price: Option<f64>,
-    pub stop_loss_price: Option<f64>,
+    pub current_price: Option<Decimal>,
+    pub target_price: Option<Decimal>,
+    pub stop_loss_price: Option<Decimal>,
     pub reason: Option<String>,
     pub expires_at: Option<String>,
 }
@@ -90,9 +95,12 @@ pub struct WatchlistItem {
 #[derive(Debug, Clone, Serialize, Deserialize, sqlx::FromRow)]
 pub struct PortfolioSnapshot {
     pub id: Option<i64>,
-    pub total_value: f64,
-    pub total_cost: f64,
-    pub total_pnl: f64,
+    #[sqlx(try_from = "f64")]
+    pub total_value: Decimal,
+    #[sqlx(try_from = "f64")]
+    pub total_cost: Decimal,
+    #[sqlx(try_from = "f64")]
+    pub total_pnl: Decimal,
     pub total_pnl_percent: f64,
     pub snapshot_date: String,
 }
@@ -100,9 +108,9 @@ pub struct PortfolioSnapshot {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PortfolioSummary {
     pub total_positions: usize,
-    pub total_value: f64,
-    pub total_cost: f64,
-    pub total_pnl: f64,
+    pub total_value: Decimal,
+    pub total_cost: Decimal,
+    pub total_pnl: Decimal,
     pub total_pnl_percent: f64,
     pub positions: Vec<PositionWithPnL>,
 }
@@ -113,11 +121,11 @@ pub struct PerformanceMetrics {
     pub winning_trades: usize,
     pub losing_trades: usize,
     pub win_rate: f64,
-    pub total_realized_pnl: f64,
-    pub average_win: f64,
-    pub average_loss: f64,
-    pub largest_win: f64,
-    pub largest_loss: f64,
+    pub total_realized_pnl: Decimal,
+    pub average_win: Decimal,
+    pub average_loss: Decimal,
+    pub largest_win: Decimal,
+    pub largest_loss: Decimal,
     pub recent_trades: Vec<Trade>,
 }
 
@@ -130,10 +138,10 @@ pub struct ActionItem {
     pub description: String,
     pub signal: String,
     pub confidence: f64,
-    pub current_price: Option<f64>,
-    pub target_price: Option<f64>,
-    pub stop_loss_price: Option<f64>,
+    pub current_price: Option<Decimal>,
+    pub target_price: Option<Decimal>,
+    pub stop_loss_price: Option<Decimal>,
     pub in_portfolio: bool,
-    pub position_pnl: Option<f64>,
+    pub position_pnl: Option<Decimal>,
     pub alert_id: Option<i64>,
 }
